+++ app/src/main/java/com/example/pocketwaifu/data/models/ChatEntity.kt
@@ -10,7 +10,7 @@ data class ChatEntity(
     @PrimaryKey(autoGenerate = true)
         val id: Int = 0,
     @ColumnInfo(name = "avatar_name")
-        val avatarId: Long,
+        val avatarId: Int,
     @ColumnInfo(name = "message_content")
         val text: String?,
     @ColumnInfo(name = "is_user")
